bts: Separate the receive handling for TBF and SBA
Make the code a bit more clear, avoid having to initialize some
variables as it is not clear if they are used or not. It would
be easier if we wouldn't have to pass ra/fn to these methods.
The early allocation of the bitvec should be okay as we expect
allocations to succeed.